fix: address PR review on bigint widening
- restore `unsigned: true` on the three bigint columns and emit
  `BIGINT UNSIGNED` in the MySQL `up()` migrations. Postgres still
  ignores `unsigned`, but MySQL keeps the non-negative invariant the
  original `int UNSIGNED` schema enforced (and that the `down()`
  preflight already assumes).
- mirror the `Number.isSafeInteger` guard in
  `bigintNumberTransformer.to()` so precision-lost values are caught at
  the write site, not later on read.
- drop `COUNT(*)::int` cast in the Postgres `down()` preflights —
  `COUNT(*)` is already bigint and the TS-side `Number(...)` handles
  the string fine; avoids a theoretical narrowing overflow on huge
  tables.
- declare `cross-env`, `unplugin-swc`, `vitest` in
  `packages/kit/package.json` devDependencies (the new `test` script
  was relying on workspace hoisting; matches the `server-db-kit`
  pattern).

+++ apps/server-core/src/adapters/database/migrations/postgres/1779354335393-Default.ts
@@ -14,13 +14,13 @@ export class Default1779354335393 implements MigrationInterface {
 
     public async down(queryRunner: QueryRunner): Promise<void> {
         const [{ overflow_count: analysisOverflow }] = await queryRunner.query(`
-            SELECT COUNT(*)::int AS overflow_count
+            SELECT COUNT(*) AS overflow_count
             FROM "analysis_entity"
             WHERE "build_size" IS NOT NULL
               AND ("build_size" > 2147483647 OR "build_size" < -2147483648)
         `);
         const [{ overflow_count: masterOverflow }] = await queryRunner.query(`
-            SELECT COUNT(*)::int AS overflow_count
+            SELECT COUNT(*) AS overflow_count
             FROM "master_images"
             WHERE "build_size" IS NOT NULL
               AND ("build_size" > 2147483647 OR "build_size" < -2147483648)
